+++ gears/processors/directives.py
@@ -6,12 +6,14 @@ from .base import BaseProcessor
 from ..asset_attributes import AssetAttributes
 from ..assets import Asset
 from ..directives_parser import DirectivesParser
+from ..exceptions import FileNotFound
 
 
 class DirectivesProcessor(BaseProcessor):
 
     def __init__(self):
         self.types = {
+            'params': self.process_params_directive,
             'require': self.process_require_directive,
             'require_directory': self.process_require_directory_directive,
             'require_tree': self.process_require_tree_directive,
@@ -37,36 +39,45 @@ class DirectivesProcessor(BaseProcessor):
             args = shlex.split(directive)
             self.types[args[0]](*args[1:])
 
-    def process_require_directive(self, path):
-        self.asset.requirements.add(self.get_asset(*self.find(path)))
+    def process_params_directive(self, *params):
+        for param in params:
+            if '=' in param:
+                key, value = param.split('=', 1)
+                self.asset.params[key] = value
 
-    def process_require_directory_directive(self, path, recursive=False):
-        path = self.get_relative_path(path, is_directory=True)
-        list = self.asset.attributes.environment.list(path, self.asset.attributes.mimetype, recursive=recursive)
+    def process_require_directive(self, path):
+        found = False
+        path = self.get_relative_path(path)
+        list = self.asset.attributes.environment.list(path, self.asset.attributes.mimetype)
         for asset_attributes, absolute_path in sorted(list, key=lambda x: x[0].path.split('/')):
             self.asset.requirements.add(self.get_asset(asset_attributes, absolute_path))
             self.asset.dependencies.add(os.path.dirname(absolute_path))
+            found = True
+        if not found:
+            raise FileNotFound(path)
+
+    def process_require_directory_directive(self, path):
+        self.process_require_directive(os.path.join(path, '*'))
 
     def process_require_tree_directive(self, path):
-        self.process_require_directory_directive(path, recursive=True)
+        self.process_require_directive(os.path.join(path, '**'))
 
     def process_require_self_directive(self):
         self.asset.requirements.add(self.asset)
 
     def process_depend_on_directive(self, path):
-        self.asset.dependencies.add(self.find(path)[1])
-
-    def find(self, require_path):
-        require_path = self.get_relative_path(require_path)
-        asset_attributes = AssetAttributes(self.asset.attributes.environment, require_path)
-        return self.asset.attributes.environment.find(asset_attributes, True)
+        found = False
+        path = self.get_relative_path(path)
+        list = self.asset.attributes.environment.list(path, self.asset.attributes.mimetype)
+        for asset_attributes, absolute_path in list:
+            self.asset.dependencies.add(absolute_path)
+            found = True
+        if not found:
+            raise FileNotFound(path)
 
     def get_relative_path(self, require_path, is_directory=False):
         require_path = os.path.join(self.asset.attributes.dirname, require_path)
-        require_path = os.path.normpath(require_path)
-        if is_directory:
-            return require_path
-        return require_path + ''.join(self.asset.attributes.extensions)
+        return os.path.normpath(require_path)
 
     def get_asset(self, asset_attributes, absolute_path):
         return Asset(asset_attributes, absolute_path, self.asset.calls)